requirements.txt
annotated-types==0.7.0
anyio==3.7.1
astroid==2.15.8
attrs==23.2.0
autopep8==2.0.1
bcrypt==4.2.1
cachetools==5.3.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coverage==7.2.1
cryptography==44.0.0
dill==0.3.8
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.103.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.5
lazy-object-proxy==1.10.0
MarkupSafe==3.0.2
mccabe==0.7.0
mypy==1.9.0
mypy-extensions==1.0.0
orjson==3.10.14
packaging==24.0
platformdirs==4.2.0
pluggy==1.4.0
psycopg2==2.9.5
pyasn1==0.6.1
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.10.5
pydantic-extra-types==2.10.1
pydantic-settings==2.7.1
pydantic_core==2.27.2
pylint==2.16.2
pytest==7.2.1
pytest-asyncio==0.20.3
pytest-mock==3.10.0
pytest-reportlog==0.4.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2023.3
PyYAML==6.0.2
requests==2.28.2
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.20
sqlalchemy-stubs==0.4
starlette==0.27.0
tomli==2.0.1
tomlkit==0.12.4
typing_extensions==4.12.2
ujson==5.10.0
urllib3==1.26.20
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.1
wrapt==1.16.0
